Database scheme is changed a lot Some repositories and entities are also changed to match new database scheme

+++ src/main/scala/com/itechart/project/domain/referee.scala
@@ -8,7 +8,7 @@ import eu.timepit.refined.types.string.NonEmptyString
 
 object referee {
 
-  final case class RefereeId(value: Long)
+  final case class RefereeId(value: Int)
 
   type RefereeFirstName = NonEmptyString
